zip with clustering stuff
this zip has the input folder with separated folders, each folder contains a random sample of video files. the prog will then do its magic and predict the class it belongs to. the input order of the files is the same so i will build a spreadsheet with the videos in order and we can map it manually if we have to.